`goto-match-paren` match paren at previous char
This is motivated by the behavior of `show-paren-mode`, but the
interaction is a bit odd.  `show-paren-mode` highlights the paren
before point, and its match.  This is nice when typing, because
after typing a close paren, point is after that character, and the
pair is highlighted.

When moving around already written code, it's not clear how to handle
consecutive parens.  `show-paren-mode` ignores a paren under point,
and highlights the open paren matching the close paren before point.
At the moment, `goto-match-paren` makes the opposite choice, matching
the paren under point.

My first choice would be to change the behavior of `show-paren-mode`.
Changing the behavior of `goto-match-paren` is easier, though, and may
be better than the current inconsistancy.