turn off `helm`
This is an experiment.  I'd like to stick with one of
(helm, ivy, ido),
but I don't know which one.

This change is mostly to let me try out the `ivy` variants of:
     - buffer selection
     - imenu
, which I use all the time

This commit comments out some things I probably want:
     - haskell-language-selections
     - notmuch-count-searches

And some I don't use much if ever:
    - dash
    - gtags
    - helm-idris