xmonad: add dual tabbed layout, remove less-used layouts
I figure topTabs is enough when I have only one window on the
workspace, and dualTabs is the same as Tall when I have only two.  I
occasionally use Full to hide the other tabs, but sending that window
to a Workspace is probably a better way.  I don't want the list of
layouts to get to long to cycle through.